refactor: delete notblank annotation

+++ src/main/java/roomescape/service/JdbcTimeService.java
@@ -34,7 +34,7 @@ public class JdbcTimeService implements TimeService {
 		}, keyHolder);
 
 		return new AddTimeResponse(
-				keyHolder.getKey(),
+				keyHolder.getKey().longValue(),
 				time
 		);
 	}
